distinguish between FileModelBundlesManager and AssetModelBundlesManager
