--- core/src/main/java/ai/doc/tensorio/core/modelbundle/AssetsModelBundleManager.java
@@ -1,4 +1,0 @@
-package ai.doc.tensorio.core.modelbundle;
-
-public class AssetsModelBundleManager {
-}
