--- core/src/main/java/ai/doc/tensorio/core/modelbundle/FilesModelBundleManager.java
@@ -1,4 +1,0 @@
-package ai.doc.tensorio.core.modelbundle;
-
-public class FilesModelBundleManager {
-}
